Organize Step Function classes in Amazon provider (#20158)
* 20139 - organize aws step_function
